pyproject.toml
[project]
name = "mcp-obsidian"
version = "0.2.1"
description = "MCP server to work with Obsidian via the remote REST plugin"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "mcp>=1.1.0",
 "pip_system_certs",
 "python-dotenv>=1.0.1",
 "requests>=2.32.3",
 "click>=8.0.0",
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.389",
    "pytest>=9.0.3",
]

[project.scripts]
mcp-obsidian = "mcp_obsidian:main"
obs = "cli_obsidian:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
